+++ Cmdlets/Get-ChildItem/Directory-with-backticks/.test.ps1
@@ -5,8 +5,8 @@ task Test-1 {
 	($r = .\Test-1.ps1)
 	if ($v610) {
 		equals $r.Count 2
-		equals $r[0] 'v6.1.0 worked'
-		equals $r[1].FullyQualifiedErrorId 'ItemNotFound,Microsoft.PowerShell.Commands.GetChildItemCommand'
+		equals $r[0].FullyQualifiedErrorId 'PathNotFound,Microsoft.PowerShell.Commands.GetChildItemCommand'
+		equals $r[1].FullyQualifiedErrorId 'PathNotFound,Microsoft.PowerShell.Commands.GetChildItemCommand'
 	}
 	else {
 		equals $r.FullyQualifiedErrorId 'PathNotFound,Microsoft.PowerShell.Commands.GetChildItemCommand'