+++ Cmdlets/Group-Object/AsHashTable/.test.ps1
@@ -1,4 +1,6 @@
 
+$v610 = $PSVersionTable.PSVersion -gt ([version]'6.0.9999')
+
 task Test-1.property.string {
 	$r = .\Test-1.property.string.ps1
 	equals $r.Count 2
@@ -10,7 +12,12 @@ task Test-2.expression.string {
 	$r = .\Test-2.expression.string.ps1
 	equals $r.Count 2
 	equals $r[0] 2
-	equals $r[1]
+	if ($v610) {
+		equals $r[1].Name README.md
+	}
+	else {
+		equals $r[1]
+	}
 }
 
 task Test-3.expression.number {