+++ Cmdlets/Group-Object/AsHashTable/Test-2.expression.string.ps1
@@ -5,5 +5,6 @@ $r = Get-ChildItem | Group-Object -Property {$_.Extension} -AsHashTable
 # 2 (.md, .ps1)
 $r.Count
 
-# $null. Expected: README.md
+# v6.1.0-preview.4 : README.md (expected)
+# older versions : $null (actual)
 $r['.md']